Update assignment submit

+++ api/courses/assignment/submit/index.php
@@ -12,24 +12,27 @@ try {
                 echo jsonResponse(400, "ค่าที่ให้มาไม่ครบหรือไม่ถูกต้อง");
                 die();
             }
-            $col = array(
-                "a_id", "u.u_firstname", "u.u_lastname", "s_datetime"
-            );
-            // $db->join("submissions_assignment s", "s.u_id=e.u_id", "LEFT");
-            // $db->join("users u", "u.u_id=e.u_id", "RIGHT");
-            // $db->where("e.c_id", $_GET["c_id"]);
-            // $db->where("e.u_role", "STUDENT");
-            // $db->where("a_id", $_GET["a_id"]);
-            $submissions = $db->rawQuery("SELECT e.u_id, s.a_id, u.u_firstname, u.u_lastname, u.u_avatar_mime_type, s_datetime FROM enrollments e LEFT JOIN
+            if (key_exists("u_id", $_GET)) {
+                $db->where('a_id', $_GET['a_id']);
+                $db->where('u_id', $_GET['u_id']);
+                $submissions = $db->getOne("submissions_assignment");
+                $submissions["s_content"] = json_decode($submissions["s_content"]);
+                echo json_encode(array("data" => $submissions));
+            } else {
+                $col = array(
+                    "a_id", "u.u_firstname", "u.u_lastname", "s_datetime"
+                );
+                $submissions = $db->rawQuery("SELECT e.u_id, s.a_id, u.u_firstname, u.u_lastname, u.u_avatar_mime_type, s_datetime FROM enrollments e LEFT JOIN
             (SELECT e.u_id, a_id, s_datetime FROM enrollments e LEFT OUTER JOIN submissions_assignment s on e.u_id=s.u_id WHERE  e.c_id=?  AND e.u_role = 'STUDENT' AND a_id=?) AS s
-            ON e.u_id=s.u_id LEFT JOIN users u ON u.u_id=e.u_id WHERE e.c_id=? AND e.u_role = 'STUDENT'", Array($_GET['c_id'], $_GET['a_id'], $_GET['c_id']));
-            
-            foreach (array_values($submissions) as $i => $obj) {
-                $submissions[$i]['u_avatar'] = !is_null($submissions[$i]['u_avatar_mime_type']);
-                unset($submissions[$i]['u_avatar_mime_type']);
+            ON e.u_id=s.u_id LEFT JOIN users u ON u.u_id=e.u_id WHERE e.c_id=? AND e.u_role = 'STUDENT'", array($_GET['c_id'], $_GET['a_id'], $_GET['c_id']));
+
+                foreach (array_values($submissions) as $i => $obj) {
+                    $submissions[$i]['u_avatar'] = !is_null($submissions[$i]['u_avatar_mime_type']);
+                    unset($submissions[$i]['u_avatar_mime_type']);
+                }
+
+                echo json_encode(array("data" => $submissions));
             }
-            
-            echo json_encode(array("data" => $submissions));
             break;
         case 'PUT': //submit
             $_PUT = json_decode(file_get_contents('php://input'), true);
@@ -40,10 +43,10 @@ try {
             $data = array(
                 "u_id" => $_SESSION["u_id"]
             );
-            foreach(array_keys($_PUT) as $key) {
+            foreach (array_keys($_PUT) as $key) {
                 $data[$key] = ($key == "s_content") ? json_encode($_PUT[$key]) : $_PUT[$key];
             }
-            echo ($db->insert("submissions_assignment", $data)) ? jsonResponse(message: "บันทีกการส่งเรียบร้อย") : jsonResponse(400, "ไม่สามารถบันทีกการส่งได้") ;
+            echo ($db->insert("submissions_assignment", $data)) ? jsonResponse(message: "บันทีกการส่งเรียบร้อย") : jsonResponse(400, "ไม่สามารถบันทีกการส่งได้");
             break;
         case "DELETE": // ลบ submits
             if (!isset($_SESSION['u_id'])) {
